requirements.txt
opencv-python~=4.10.0.82
numpy~=1.26.3
torch~=2.3.1+cu118
pillow~=10.2.0
tqdm~=4.66.4
torchvision~=0.18.1+cu118